### 2021-02-11 #428 @dbarneche
`dataaimsr`: Australian Institute of Marine Science (AIMS) Data Platform API Client which provides easy access to AIMS Data Platform scientific data and information
Submitting Author: Diego R. Barneche (@dbarneche)
Other Authors: AIMS Datacentre team: (Jeffrey L. Sheehan, Greg Coleman, Mark Rehbein)
Repository: https://github.com/open-AIMS/dataaimsr
Version submitted: 1.0.2
Reviewers:  @boshek
Reviewers:  @techisdead
Version accepted: TBD

---

```
Package: dataaimsr
Title: AIMS Data Platform API Client
Version: 1.0.2
Authors@R: c(person("Diego R.", "Barneche", email = "d.barneche@aims.gov.au", role = c("aut", "cre")), person("AIMS Datacentre", role = "aut", email = "adc@aims.gov.au"))
Description: AIMS Data Platform API Client which provides easy access to AIMS Data Platform scientific data and information.
    httr,
    parsedate
    plyr,
    tidyverse,
    ggmap,
```

`dataaimsr` is an R package written to provide open access to decades of field measurements of atmospheric and oceanographic parameters around the coast of Australia, conducted by the [Australian Institute of Marine Science]( https://www.aims.gov.au/) (AIMS). The package communicates with the recently-developed AIMS Data Platform API via an API key.

The data retrieval functions within `dataaimsr` are designed to be useful to any researcher interested in long-term, instrument-based marine environmental data from tropical Australia.

No because the datasets have been always maintained internally at AIMS. This is the first attempt of the institute to make it broadly available via an R package.

N/A

N/A
