### 2021-02-15 #428 @ldecicco-USGS
Hi @dbarneche ! Thanks for your submission. I'll be the editor for this process. The "fit" for the package and rOpenSci looks great, a very suitable package.

The first step is the "editor checks". Once this is completed, I'll be asking around for reviewers.

## Editor checks:

---

#### Editor comments

The standard `devtools::check()` resulted in:
```r
> checking package dependencies ... NOTE
  Suggests orphaned package: 'ggmap'

> checking for unstated dependencies in vignettes ... NOTE
  '::' or ':::' imports not declared from:
    'dplyr' 'tidyr'

0 errors v | 0 warnings v | 2 notes x
```

A standard tool we use is `goodpractice::gp()`, which resulted in:
```r
Preparing: rcmdcheck
── GP dataaimsr ────────────────────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all
    package code in general. 80% of code lines are covered
    by test cases.

    R/aims_data.R:169:NA
    R/aims_data.R:170:NA
    R/aims_data.R:171:NA
    R/aims_data.R:176:NA
    R/aims_data.R:177:NA
    ... and 22 more lines

  ✖ avoid long code lines, it is bad for
    readability. Also, many people prefer editor windows
    that are about 80 characters wide. Try make your lines
    shorter than 80 characters

    R\aims_data.R:29:1

  ✖ fix this R CMD check NOTE: '::' or ':::'
    imports not declared from: 'dplyr' 'tidyr'
```

Coverage looks pretty good:
```r
covr::package_coverage()
dataaimsr Coverage: 80.00%
R/zzz.R: 0.00%
R/aims_data.R: 80.00%
R/page_data.R: 88.24%
R/helpers.R: 90.32%
R/filter_values.R: 90.91%
R/aims_data_doi.R: 100.00%
R/expose_attributes.R: 100.00%
```

You suggest `tidyverse`, which is typically not recommended:
https://www.tidyverse.org/blog/2018/06/tidyverse-not-for-packages/

Because some package installations will install all suggested packages, suggesting `tidyverse` could be a major undertaking for your users. Glancing though your vignettes, it looks like you could suggest just `dplyr` and `tidyr`, and probably `purrr` if you consider dropping the `plyr::llply` suggestion and replace it with `purrr:mapdf`.

On that note, `plyr` is a package you'll probably want to avoid if possible. If you are already steering your users towards tidyverse, I think it's not a stretch to drop the `plyr` suggestion and move to a `purrr` alternative.

Many of the help file (.Rd generated from roxygen) don't have working examples, which is always helpful for users.

Due to the nature of the package, you may want to have a look here for additional testing strategies:
https://books.ropensci.org/http-testing/index.html

So with that being said, I'll start the process for finding reviewers, but I'd encourage you to consider the above comments in the meantime.

You can add the rOpenSci 'under review' badge to your README:

Full link should be:
