### 2021-02-15 #428 @dbarneche
Hi @ldecicco-USGS ! Thank you very much for such a prompt and thorough feedback. I'll start working on your suggestions today.

I have a parallel question in the meantime as well, if you don't mind. We are hoping to submit a paper to JOSS about our package once (and if) it is properly approved by rOpenSci. Is it OK for us to keep working on the paper while the package is being reviewed here?

Also, the data team at AIMS is also developing a web-based data explorer for the same datasets. It serves as a complementary tool for users who prefer a more interactive form of navigation/exploration. One of the features of the tool is to export an R script based on dataaimsr to download the data type/window chosen by the user. There are other features such as downloading the data in csv or a printing the time series as png.

This is my first time submitting a package to rOpenSci and a paper to JOSS. In your opinion, should we add a section about this complementary tool to the paper, or should it be another standalone article?
