### 2021-02-16 f1bb2ba
removed plyr and tidyverse from suggestions and vignettes
