+++ man/find_api_key.Rd
@@ -32,7 +32,7 @@ the API Key automatically.
 There are two types of data currently available through the
 \href{https://open-aims.github.io/data-platform}{AIMS Data Platform API}:
 \href{https://weather.aims.gov.au/#/overview}{Weather} and
-\href{https://apps.aims.gov.au/metadata/view/4a12a8c0-c573-11dc-b99b-00008a07204e}{Sea Water Temperature Loggers}.
+\href{https://tinyurl.com/h93mcojk}{Sea Water Temperature Loggers}.
 They are searched internally via unique DOI identifiers which
 can be obtained by the function \code{\link{aims_data_doi}} (see Examples).
 Only one DOI at a time can be passed to the argument \code{doi}.
