### 2021-02-16 dff2d8e
tests for all functions
