+++ tests/testthat/setup.R
@@ -8,7 +8,7 @@ add_attr <- function(x, ...) {
 weather_doi <- aims_data_doi("weather")
 ssts_doi <- aims_data_doi("temp_loggers")
 wrong_doi <- "string"
-my_api_key <- "dFpi9Qum8Q8cNAxnK4Q8Q4PoRrd7SBp9asHISDnI"
+my_api_key <- "1NIik2HDo36shE3FG3MfFMLlNBfEOL945FIH7fX0"
 w_filters <- list(series_id = 64, from_date = "2018-01-01",
                   thru_date = "2018-01-10")
 w_filters_b <- add_attr(w_filters, size = 1e4)
