+++ pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 2.7.3
 pkgdown: 1.6.1.9000
-pkgdown_sha: c931fe24c135c6ca704ae9f300470e9c66d8bfe7
+pkgdown_sha: 1c28ef9d9476e35974e1933de76a9aaa24d4cd27
 articles:
   navigating: navigating.html
   tloggers: tloggers.html
   weather: weather.html
-last_built: 2021-02-11T07:36Z
+last_built: 2021-02-16T00:52Z
 
