### 2021-02-17 #428 @karthik
> Is it OK for us to keep working on the paper while the package is being reviewed here?

Yes very much so. The paper will evolve further in the JOSS review.

> In your opinion, should we add a section about this complementary tool to the paper, or should it be another standalone article?

You can certainly mention it. A web based explorer for data does not qualify as research software under JOSS so you can't submit that as a separate paper anyway.
