### 2021-03-02 #428 @boshek
@ldecicco-USGS here is my review of the `dataaimsr`. Overall a really great package that accesses a valuable data source. I have tried to summarize my key topics below. I think the biggest two point for me are:

1. Spend some additional time to make sure the API keys and the presence of an internet connections are checked for both users and for CI/CRAN services. This will help you debug future issues and provides some nice guardrails for users.
2. I'd advocate strongly that `aims_data` reliably returns a data.frame and exposes those other elements (citation etc) as attributes.

Great job with this package!

- [ ] **Examples** (that run successfully locally) for all exported functions

>##### For packages co-submitting to JOSS
>
>- [X] The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
>
>- [X] **A short summary** describing the high-level functionality of the software
>- [X] **Authors:**  A list of authors with their affiliations
>- [X] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

Estimated hours spent reviewing: 8.

---

# Review Comments

- **Statement of Need**: I think that the README could actually be made _shorter_. I think it would be helpful to outline why a user would want to use R to interact with this data. To me the specific value add here is that you are connecting the AIMS api with all the amazing capabilities of R. I think the text in the _Overview_ section is valuable. I just wonder if it could be condensed a little to get users to specifically to the installation instructions as really what they are after here is the software itself.
- **Installation instructions**: I would revamp the installation instructions so that getting the API key is explictly a step in the installation instructions. I would also include information about how the API key should be kept secret. You might consider linking to [this](https://cran.r-project.org/web/packages/httr/vignettes/secrets.html) article.
- **Vignette(s)**: See comment in API keys section. I don't think ggmap is a particularly useful here. Given ggmaps' own API requirements I would strongly advise against including it here as a dependency and including it in the vignettes. If you really want a map, I would suggest try an sf approach. But I think there are plenty of tutorials around to tackle mapping in R and it detracts from the core focus of dataaimsr.
- **Examples**: I think the example in `aims_data` could be significantly shortened. The examples section is meant to be illustrative rather than comprehensive. But I wouldn't waste all that great code. This is an opinion so not a requirement for acceptance.  I would simply move it to a vignette. I am a bit uncertain how you are going to manage with CRAN, the requirement for the API keys and the ability to run examples. Here I think you might need to build in some ability for the function to 'fail gracefully' when the API key isn't available. That is, don't error but instead return an informative message. That same applies to checking if someone has a) internet and b) if the API is down for whatever reason. Again this is particularly relevant if CRAN submission is desired. I think you can wrap them in a `try` function. That might work.
**Community guidelines**: Missing a CONTRIBUTING guide _and_ (though not required) a CODE_OF_CONDUCT file.

- I think it is worth adopting stronger language about the APi keys in the paper to ensure that people don't expose those publicly.
- I would modify the map presented in the paper. This isn't a scientific map. Either I would remove it or I would provide something that use open data. Something like this roughed up example would work:

```r
library(dataaimsr)
library(sf)
library(rnaturalearth)
library(tmap)

# see ?aims_data_doi for argument names
ssts_doi <- aims_data_doi("temp_loggers")
sdata <- aims_data(ssts_doi, api_key = NULL,
                   summary = "summary-by-series")

oceania_boundary <- ne_countries(continent = "oceania", returnclass = "sf") %>%
  st_transform(3112)

sdata_sf <- sdata %>%
  dplyr::mutate(cols = cal_obs * 1e-3) %>%
  tidyr::drop_na(lon, lat) %>%
  st_as_sf(coords = c("lon", "lat"), crs = 4326) %>%
  st_transform(3112)

tm_shape(oceania_boundary) +
  tm_polygons() +
  tm_shape(sdata_sf) +
  tm_dots(col = 'cols', size = 0.1)
```

- I think that your R dependency could be 3.3.0 rather than 3.6.0.
- I am wondering about the usefulness of the `aims_data_doi` function. Would it not be more straight forward for a user to simply to expose the `target` argument directly to `aims_data` and other functions like `expose_attributes`?
- I would suggest that adding `aims_*` prefixes to all exposed functions would be a good idea.
- I think you reconsider exposing most of the functions that handle http requests. That would things like `handle_error` and `page_data`. Those all can be internal functions and don't need to be user facing.
- If the object returned by `aims_data` was always just a data.frame I think that would be beneficial. Instead of returning a list I think you could stored those other elements as attributes attached to the object and then construct extract functions. So something like `aims_citation` which takes the object returned by `aims_data` and return the citation. Typically returning objects of different classes depending on the value of an argument is considered not best practice.
- I think this approach could be extended to parameters
- I don't consider this required for acceptance but I would highly recommend considering adding classes to the `aims_data` object and implementing `print` and `plot` methods.
- Add something like `curl::has_internet()` to fail "gracefully" when lacking an internet connection.
- I think the README could benefit from a short usage example.
- In the DESCRIPTION I think the url could also include a link to the API. They just need to be separated by a comma.

## API keys

- I think that it would be worthwhile in include a check for the env variable when the package is loaded. Using `find_api_key` would probably work fine.
- Exclude the actual API from the RMD's. There are few ways around actually needing to include it. You could precompile your vignettes like this:
  - https://github.com/bcgov/bcdata/blob/master/vignettes/precompile.R
  - Or you can set an phantom environmental variable which results in nothing actually being run on CRAN: https://github.com/ropensci/tidyhydat/blob/2c177aec31d5e8fe296b5336b786b01bb83841c9/vignettes/tidyhydat_an_introduction.Rmd#L18
- For testing, I think the best approach if you want to test directly on the API is to add a token as GitHub secret. However ropensci does have a suite mocking tools that facilitate exactly this type of thing. For example: https://github.com/ropensci/webmockr
