### 2021-03-03 #428 @techisdead
Hey - just jumping in here to say I'd be happy to review. I'm not a regular ROpenSci contributor but an everyday programmer/builder/reviewer of R packages at @SymbolixAU and in Australia (and very interested in ecological open data).

I can't look at it for a couple of days but will put some time aside on the weekend, if that's cool.
