### 2021-03-04 #428 @ldecicco-USGS
@techisdead , thanks for volunteering!

Our [reviewers guide](https://devguide.ropensci.org/reviewerguide.html) details what we look for in a package review and includes links to example reviews. Our standards are detailed in our [packaging guide](https://devguide.ropensci.org/building.html), and we provide [reviewer](https://devguide.ropensci.org/reviewtemplate.html) template for you to use. Please make sure you do not have a [conflict of interest](https://devguide.ropensci.org/policies.html#coi) preventing you from reviewing this package. If you have questions or feedback, feel free to let me know.

We ask reviewers to complete reviews in three weeks, so if you were to finish a review this weekend, that would be fantastic.
