### 2021-03-14 #428 @techisdead
- [x] **Examples** (that run successfully locally) for all exported functions

>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

Great package and data sets.  I'm excited to see how the API evolves and think it's great this package is being developed alongside.

Overall I agree with the observations made by @ldecicco-USGS and @boshek (exept I liked the long README :-) )

I found the package well documented and easy to use on the whole. I ran into an issue with tests that hadn't been noted previously - perhaps something has changed on the server side ? Happy to help figure it out.

More details follow:

* Tests:  I agree with @boshek about not exposing an api key in your tests.  I've used {httptest} to mock an api locally and it works well, but using github secrets or {webmockr} would be good choices too

Also `devtools::test()` failed for me.

```
> devtools::test()
Loading dataaimsr
Testing dataaimsr
✓ |  OK F W S | Context
x |  30 4     | aims_data [6.1 s]
─────────────────────────────────────────────────────────
Failure (test-aims_data.R:16:3): Correct structure
all(c("metadata", "citation", "data") %in% names(check_sa)) is not TRUE

`actual`:   FALSE
`expected`: TRUE

Failure (test-aims_data.R:22:3): Correct structure
all(c("metadata", "citation", "data") %in% names(check_sb)) is not TRUE

`actual`:   FALSE
`expected`: TRUE

Failure (test-aims_data.R:41:3): Wrong filters
all(c("metadata", "citation", "data") %in% names(check_sc)) is not TRUE

`actual`:   FALSE
`expected`: TRUE

Failure (test-aims_data.R:45:3): Wrong filters
all(c("metadata", "citation", "data") %in% names(check_sd)) is not TRUE

`actual`:   FALSE
`expected`: TRUE

```

This is because the structure returnes from the two endpoints is different

`aims_data(weather_doi, ...)` returns data of the form

```
> str(check_wa)
List of 3
 $ metadata: chr "Metadata record https://doi.org/10.25845/5c09bf93f315d"
 $ citation: chr "Australian Institute of Marine Science (AIMS). 2009, Australian Institute of Marine Science Automatic Weather S"| __truncated__
 $ data    :'data.frame':	1297 obs. of  14 variables:

 ...
```

```
> names(check_wa)
[1] "metadata" "citation" "data"

```

Whereas `citation` is nested under `$metadata` for `aims_data(ssts_doi, ... )`

```

> str(check_sa)
List of 3
 $ metadata:List of 2
  ..$ metadata: chr "Metadata record https://doi.org/10.25845/5b4eb0f9bb848"
  ..$ citation: chr "Australian Institute of Marine Science (AIMS). 2017, AIMS Sea Temperature Observing System (AIMS Temperature Lo"| __truncated__
 $ NA      : NULL
 $ data    :'data.frame':	433 obs. of  22 variables:

 ...

```

```
> names(check_sa)
[1] "metadata" NA         "data"

```

Has something changed in the underlying API?

This impact the `page_data` tests too

```

───────────────────────────────────────────────────────────────────
✓ |   6       | aims_data_doi
✓ |   7       | expose_attributes
✓ |  12       | filter_values [0.9 s]
✓ |  39       | helpers [0.5 s]
x |  29 5     | page_data [4.7 s]
──────────────────────────────────────────────────────────────────────
Failure (test-page_data.R:22:3): Correct structure
`check_sa` has length 4, not length 5.

Failure (test-page_data.R:30:3): Correct structure
`check_sb` has length 2, not length 3.

Failure (test-page_data.R:31:3): Correct structure
all(c("metadata", "citation", "data") %in% names(check_sb)) is not TRUE

`actual`:   FALSE
`expected`: TRUE

Failure (test-page_data.R:36:3): Correct structure
`check_sa_n` has length 2, not length 3.

Failure (test-page_data.R:37:3): Correct structure
all(c("metadata", "citation", "data") %in% names(check_sa_n)) is not TRUE

`actual`:   FALSE
`expected`: TRUE
─────────────────────────────────────────────────────────────────────

══ Results ═══════════════════════════════════════════════════════════

```

* **data structures** strongly agree with @boshek about restructuring `aims_data` to returns a `data.frame` with attributes.  Then also having `aims_citation` and `aims_metadata` as well to extract those attributes individually.
