+++ tests/testthat/setup.R
@@ -5,10 +5,7 @@ add_attr <- function(x, ...) {
   c(x, add_args)
 }
 
-weather_doi <- aims_data_doi("weather")
-ssts_doi <- aims_data_doi("temp_loggers")
-wrong_doi <- "string"
-my_api_key <- "1NIik2HDo36shE3FG3MfFMLlNBfEOL945FIH7fX0"
+my_api_key <- find_api_key(NULL)
 w_filters <- list(series_id = 64, from_date = "2018-01-01",
                   thru_date = "2018-01-10")
 w_filters_b <- add_attr(w_filters, size = 1e4)
