+++ R/helpers.R
@@ -92,7 +92,7 @@ make_base_end_pt <- function(doi, aims_version = NA) {
 #'
 #' When \code{\link[jsonlite]{fromJSON}} returns a list, format list names
 #'
-#' @inheritParams aims_data
+#' @inheritParams page_data
 #'
 #' @param results A \code{\link[jsonlite]{fromJSON}} list
 #' generated by \code{\link{json_results}}.
@@ -139,4 +139,31 @@ find_api_key <- function(api_key) {
   } else {
     api_key
   }
+}
+
+#' AIMS Dataset DOI retriever
+#'
+#' Returns DOI for a given dataset
+#'
+#' @param target A \code{\link[base]{character}} vector of length 1 specifying
+#' the dataset. Only \code{weather} or \code{temp_loggers} are currently
+#' allowed.
+#'
+#' @return A \code{\link[base]{character}} vector
+#' containing the dataset DOI string.
+#'
+#' @author AIMS Datacentre \email{adc@aims.gov.au}
+#'
+#' @examples
+#' \dontrun{
+#' library(dataaimsr)
+#' weather_doi <- data_doi("weather")
+#' ssts_doi <- data_doi("temp_loggers")
+#' }
+data_doi <- function(target) {
+  if (!(target %in% c("weather", "temp_loggers"))) {
+    stop("Wrong type of data target, only \"weather\"",
+         "or \"temp_loggers\" are allowed")
+  }
+  getOption(paste0("dataaimsr.", target))
 }
