### 2021-03-17 c4a26fb
updated DESCRIPTION by changing R dependency to 3.3.0; added API and Key request links; and added editor and refs to the reviewers list
