+++ DESCRIPTION
@@ -2,10 +2,10 @@ Package: dataaimsr
 Type: Package
 Title: AIMS Data Platform API Client
 Version: 1.0.2
-Authors@R: c(person("Diego R.", "Barneche", email = "d.barneche@aims.gov.au", role = c("aut", "cre")), person("AIMS Datacentre", role = "aut", email = "adc@aims.gov.au"))
+Authors@R: c(person("Diego R.", "Barneche", email = "d.barneche@aims.gov.au", role = c("aut", "cre")), person("AIMS Datacentre", role = "aut", email = "adc@aims.gov.au"), person("Sam", "Albers", role = "rev", email = "sam.albers@gov.bc.ca"), person("Elizabeth", "Stark", role = "rev", email = "estark@symbolix.com.au"), person("Laura", "DeCicco", role = "rev", email = "ldecicco@usgs.gov"))
 Description: AIMS Data Platform API Client which provides easy access to AIMS Data Platform scientific data and information.
 Depends:
-    R (>= 3.6.0)
+    R (>= 3.3.0)
 Imports:
     httr,
     jsonlite,
@@ -23,7 +23,7 @@ License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
 VignetteBuilder: knitr
-URL: https://open-aims.github.io/dataaimsr
+URL: https://open-aims.github.io/dataaimsr, https://open-aims.github.io/data-platform/key-request, https://open-aims.github.io/data-platform/swagger/
 BugReports: https://github.com/open-aims/dataaimsr/issues
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.1.1
