### 2021-03-18 3ead676
restructured the output of aims_data, implemented new class, and allow it to fail gracefully
