### 2021-03-18 799e4c2
Added convenience functions to extract most important attributes from aims_data output
