+++ R/aimsdf-methods.R
@@ -1,0 +1,248 @@
+#' summary.aimsdf
+#'
+#' @param object An object of class \code{\link{aimsdf}} as
+#' returned by \code{\link{aims_data}}.
+#' @param ... Unused.
+#'
+#' @return A list containing summary info from the input data.frame.
+#'
+#' @export
+summary.aimsdf <- function(object, ...) {
+  x <- list(
+    cit = aims_citation(object),
+    met = aims_metadata(object),
+    pars = aims_parameters(object),
+    dim = dim(object),
+    type = attr(object, "type"),
+    target = attr(object, "target"),
+    hdf = head(data.frame(object))
+  )
+  target <- make_pretty_data_label(x$target)
+  if (grepl("summary", x$type)) {
+    x$type <- capitalise(x$type)
+    cat(paste0(x$type, " from the ", target, " dataset"), sep = "")
+    target
+    cat("\n\n")
+    cat("A data.frame with ", x$dim[1], " observations and ", x$dim[2],
+        " variables:\n", sep = "")
+    print(x$hdf)
+    cat("...\n")
+  } else {
+    cat(paste0(target,
+               " ", x$type, " data containing the following attributes:\n"),
+        paste0("  - citation: ", x$cit, "\n"),
+        paste0("  - metadata: ", x$met, "\n"),
+        paste0("  - parameters: ", paste0(x$pars, collapse = "; ")),
+        sep = "")
+    cat("\n\n")
+    cat("A data.frame with ", x$dim[1], " observations and ", x$dim[2],
+        " variables:\n", sep = "")
+    print(x$hdf)
+    cat("...\n")
+  }
+  invisible(x)
+}
+
+#' print.aimsdf
+#'
+#' @param x An object of class \code{\link{aimsdf}} as
+#' returned by \code{\link{aims_data}}.
+#' @param ... Not used.
+#'
+#' @return A list containing a summary of the model fit as returned a
+#' brmsfit for each model.
+#'
+#' @export
+print.aimsdf <- function(x, ...) {
+  NextMethod()
+}
+
+#' head.aimsdf
+#'
+#' @param x An object of class \code{\link{aimsdf}} as
+#' returned by \code{\link{aims_data}}.
+#' @param ... Not used.
+#'
+#' @export
+head.aimsdf <- function(x, ...) {
+  NextMethod()
+}
+
+#' tail.aimsdf
+#'
+#' @param x An object of class \code{\link{aimsdf}} as
+#' returned by \code{\link{aims_data}}.
+#' @param ... Not used.
+#'
+#' @export
+tail.aimsdf <- function(x, ...) {
+  NextMethod()
+}
+
+#' plot.aimsdf
+#'
+#' Plotting options for aimsdf objects
+#'
+#' @param x An object of class \code{\link{aimsdf}} as
+#' returned by \code{\link{aims_data}}.
+#' @param ptype Type of plot. Can either be "time_series" or "map".
+#' @param pars Which parameters to plot? Only relevant if ptype is
+#' "time_series"
+#' @param ... Not used.
+#'
+#' @details Currently plots cannot be customised. Summary datasets can only
+#' be represented by maps.
+#'
+#' @return An object of class \code{\link[ggplot2]{ggplot}}.
+#'
+#' @importFrom rnaturalearth ne_countries
+#' @importFrom dplyr %>% mutate filter group_by summarise
+#' @importFrom tidyr drop_na
+#' @importFrom sf st_transform st_as_sf
+#' @importFrom ggplot2 ggplot geom_sf theme_classic theme element_rect labs
+#' @importFrom ggplot2 aes scale_colour_distiller guides guide_colourbar
[…]
