### 2021-03-18 1d0c8de
updated tests for aims_data
