### 2021-03-18 f11b5d3
allow filter_values to fail gracefully without proper internet connection
