### 2021-03-18 557f41a
updated vignettes based on recent changes to aims_data, removed keys, and simplified package dependencies
