+++ vignettes/navigating.Rmd
@@ -15,35 +15,19 @@ been downloaded.
 [0]: https://www.aims.gov.au/
 [1]: https://open-aims.github.io/data-platform/key-request
 
-Once you have obtained your API Key, you can either place it 
-permanently on your `.Renviron` file and set the object `my_api_key` to 
-`NULL` in the chunk below, or if for some reason you are having 
-difficulty placing your API Key permanently to the `.Renviron`, then 
-just paste it to the object `my_api_key` below.
+As per the installation instructions, we strongly suggest that you hide your
+API Key permanently in your `.Renviron` file and set the object `my_api_key` to
+`NULL` in the chunk below. You can read more about why that is important
+[here](https://CRAN.R-project.org/package=httr/vignettes/secrets.html).
 
-```{r, echo = FALSE, message = FALSE, warning = FALSE}
-library(dplyr)
-library(tidyr)
-library(ggplot2)
-library(ggmap)
-library(dataaimsr)
-my_api_key <- "1NIik2HDo36shE3FG3MfFMLlNBfEOL945FIH7fX0"
+```{r}
+# set my_api_key to NULL after successfully placing it in .Renviron
+my_api_key <- NULL
 ```
 
-```{r, eval = FALSE}
-# set my_api_key to NULL if successfully placed in .Renviron
-# paste your key where it says api-key-for-r-notebook
-my_api_key <- "api-key-for-r-notebook"
-```
+We now load `dataaimsr`:
 
-Then install `dataaimsr` following the documentation on our
-[README](../index.html) page. Once that's out of the way, we load the packages needed for this vignette:
-
-```{r, eval = FALSE}
-library(dplyr)
-library(tidyr)
-library(ggplot2)
-library(ggmap)
+```{r}
 library(dataaimsr)
 ```
 
@@ -111,7 +95,7 @@ The `summary` argument here is key. It should only be flagged when the
 user wants an overview of the available data. Again, this currently
 implemented for the Sea Water Temperature Loggers dataset. One can
 visualise `summary-by-series` or `summary-by-deployment`. The output of
-`aims_data` when summary is `NA` (the default) is a `data.frame`.
+`aims_data` is a `data.frame` of class `aimsdf`.
 
 Notice that `sdata` contains a lot of information, most of which is
 related to site / series / parameter ID. Each row corresponds to a
@@ -126,35 +110,24 @@ hand.
 Also note that there are three columns containing the total number of 
 observations in a series: `uncal_obs`, `cal_obs` and `qc_obs`, which 
 respectively stand for uncalibrated, calibrated, and quality-controlled 
-observations. Calibrated and quality-controlled are generally the same,
-and currently there is no public information on AIMS quality-control
-methods and algorithms. However, if this is crucial information to you,
-please reach out to the
-[Data Manager from the AIMS Data Centre](adc@aims.gov.au). So let's
-go ahead and plot these data on a map of Australia, while colouring
-based on the total amount of calibrated observations:
+observations. Calibrated and quality-controlled are generally the same.
+Instruments are routinely calibrated (mostly once a year) in a temperature-controlled water bath and corrections applied to the data. After
+calibration, all data records are quality controlled based on the following
+tests: 1) clip to in-water only data, using deployment's metadata, 2)
+impossible value check: data outside a fixed temperature range (14˚C – 40˚C)
+is flagged as bad data, 3) spike test: individual extreme values are flagged
+as probably bad and 4) Excessive gradient test: pairs of data that present a
+sudden change in the slope are flagged as probably bad. If any data record
+fails at least one of the tests, a QC flag equal to 2 is returned, otherwise,
+the QC flag is set to 1.
+
[…]
