+++ vignettes/tloggers.Rmd
@@ -15,25 +15,22 @@ datasets. This vignette assumes you have obtained an
 [1]: https://open-aims.github.io/dataaimsr/articles/navigating.html
 [2]: https://open-aims.github.io/data-platform/key-request
 
-Let's start by loading some packages that we are going to need down the track,
-and store the API key as an object---if you successfully placed your API Key
-permanently to the `.Renviron`, then just set the object `my_api_key` to `NULL`
-in the chunk below:
+As per the installation instructions, we strongly suggest that you hide your
+API Key permanently in your `.Renviron` file and set the object `my_api_key` to
+`NULL` in the chunk below. You can read more about why that is important
+[here](https://CRAN.R-project.org/package=httr/vignettes/secrets.html).
 
-```{r, eval = FALSE}
-library(purrr)
-library(ggplot2)
-library(dataaimsr)
-# set my_api_key to NULL if successfully placed in .Renviron
-# paste your key where it says api-key-for-r-notebook
-my_api_key <- "api-key-for-r-notebook"
+```{r}
+# set my_api_key to NULL after successfully placing it in .Renviron
+my_api_key <- NULL
 ```
 
-```{r, echo = FALSE, message = FALSE, warning = FALSE}
+Let's start by loading the packages needed for this vignette:
+
+```{r}
 library(purrr)
-library(ggplot2)
 library(dataaimsr)
-my_api_key <- "1NIik2HDo36shE3FG3MfFMLlNBfEOL945FIH7fX0"
+library(ggplot2)
 ```
 
 ## Discovering the dataset
@@ -94,24 +91,13 @@ ddata <- aims_data("temp_loggers", api_key = my_api_key,
 head(ddata)
 ```
 
-Notice that `sdata$data` contains a lot of information, most of which is
+Notice that `sdata` contains a lot of information, most of which is
 related to site / series / parameter ID. Each row corresponds to a
 unique series. The columns `time_coverage_start` and `time_coverage_end` are
 probably one of the most valuable pieces of information. They provide the user
 with the window of data collection for a particular series, which is probably
 crucial to decide whether that particular series is of relevance to the
 specific question in hand.
-
-Also note that there are three columns containing the total number of 
-observations in a series: `uncal_obs`, `cal_obs` and `qc_obs`, which 
-respectively stand for uncalibrated, calibrated, and quality-controlled 
-observations. Calibrated and quality-controlled are generally the same,
-and currently there is no public information on AIMS quality-control
-methods and algorithms. However, if this is crucial information to you,
-please reach out to the
-[Data Manager from the AIMS Data Centre](adc@aims.gov.au). So let's
-go ahead and plot these data on a map of Australia, while colouring
-based on the total amount of calibrated observations:
 
 The benefits to choosing a data `series` (or the numeric equivalent,
 `series_id`) is that it comes from one location and parameter type (here only
@@ -144,10 +130,16 @@ agincourt <- aims_data("temp_loggers", api_key = my_api_key,
 We can check that the query filters worked:
 
 ```{r}
-range(agincourt$data$time)
+range(agincourt$time)
 ```
 
-We can even visually compare multiple series at once. For instance, let's
+We can then visualise where in Australia that data is placed:
+
+```{r, message = FALSE, warning = FALSE, fig.width = 5, fig.height = 5}
+plot(agincourt, ptype = "map")
+```
+
+We can also visually compare multiple series at once. For instance, let's
 compare the air temperature data from Davies Reef and Bramble Cay for the
 same period of time:
 
[…]

[…]
