### 2021-03-18 c4cab26
updated function and package documentation
