+++ man/plot.aimsdf.Rd
@@ -1,0 +1,44 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/aimsdf-methods.R
+\name{plot.aimsdf}
+\alias{plot.aimsdf}
+\title{plot.aimsdf}
+\usage{
+\method{plot}{aimsdf}(x, ..., ptype, pars)
+}
+\arguments{
+\item{x}{An object of class \code{\link{aimsdf}} as
+returned by \code{\link{aims_data}}.}
+
+\item{...}{Not used.}
+
+\item{ptype}{Type of plot. Can either be "time_series" or "map".}
+
+\item{pars}{Which parameters to plot? Only relevant if ptype is
+"time_series"}
+}
+\value{
+An object of class \code{\link[ggplot2]{ggplot}}.
+}
+\description{
+Plotting options for aimsdf objects
+}
+\details{
+Currently plots cannot be customised. Summary datasets can only
+be represented by maps.
+}
+\examples{
+\dontrun{
+library(dataaimsr)
+wdf <- aims_data("weather", api_key = NULL,
+                 filters = list(site = "Yongala",
+                                from_date = "2018-01-01",
+                                thru_date = "2018-01-02"))
+plot(wdf, ptype = "map")
+plot(wdf, ptype = "time_series")
+# summary datasets can only return maps
+sdf <- aims_data("temp_loggers", api_key = NULL,
+                 summary = "summary-by-deployment")
+plot(sdf, ptype = "map")
+}
+}
