+++ man/aims_data.Rd
@@ -18,19 +18,27 @@ filters for the data query (see Details)}
 functions \code{\link{page_data}} and \code{\link{next_page_data}}}
 }
 \value{
-If \code{summary} is passed as an additional argument (see code
-example below), \code{aims_data} returns a \code{\link[base]{data.frame}}
-containing the summary information for the dataset (NB: currently,
-\code{summary} only works for the temperature logger dataset).
+\code{aims_data} returns a \code{\link[base]{data.frame}} of class
+\code{\link{aimsdf}}.
 
-Otherwise, if actual data is downloaded, a \code{\link[jsonlite]{fromJSON}}
-\code{\link[base]{list}} containing three elements:
+If \code{summary} is passed as an additional argument,
+the output shows the summary information for the target dataset (i.e.
+weather or temperature loggers)
+(NB: currently, \code{summary} only works for the temperature logger
+database). If \code{summary} is \emph{not} passed as an additional argument, then
+the output contains monitoring data.
+The output also contains five attributes (empty strings if
+\code{summary} is passed as an additional argument):
 \itemize{
 \item{\code{metadata}}{a \href{https://www.doi.org/}{DOI} link
 containing the metadata record for the data series.}
 \item{\code{citation}}{the citation information for the particular
 dataset.}
-\item{\code{data}}{an output \code{\link[base]{data.frame}}.}
+\item{\code{parameters}}{The measured parameters comprised in the
+output.}
+\item{\code{type}}{The type of dataset. Either "monitoring" if
+\code{summary} is not specified, or a "summary-by-" otherwise.}
+\item{\code{target}}{The input target.}
 }
 }
 \description{
@@ -85,7 +93,7 @@ library(dataaimsr)
 wdf_a <- aims_data("weather", api_key = NULL,
                    filters = list(site = "Yongala",
                                   from_date = "2018-01-01",
-                                  thru_date = "2018-01-02"))$data
+                                  thru_date = "2018-01-02"))
 
 # 2. downloads weather data from all sites
 # under series_id 64 from Davies Reef
@@ -93,7 +101,7 @@ wdf_a <- aims_data("weather", api_key = NULL,
 wdf_b <- aims_data("weather", api_key = NULL,
                    filters = list(series_id = 64,
                                   from_date = "1991-10-18",
-                                  thru_date = "1991-10-19"))$data
+                                  thru_date = "1991-10-19"))
 head(wdf_b)
 range(wdf_b$time)
 
@@ -103,7 +111,7 @@ range(wdf_b$time)
 wdf_c <- aims_data("weather", api_key = NULL,
                    filters = list(series_id = 64,
                                   from_date = "1991-10-18T06:00:00",
-                                  thru_date = "1991-10-18T12:00:00"))$data
+                                  thru_date = "1991-10-18T12:00:00"))
 head(wdf_c)
 range(wdf_c$time)
 
@@ -111,7 +119,7 @@ range(wdf_c$time)
 # within a defined date range
 wdf_d <- aims_data("weather", api_key = NULL,
                    filters = list(from_date = "2003-01-01",
-                                  thru_date = "2003-01-02"))$data
+                                  thru_date = "2003-01-02"))
 # note that there are multiple sites and series
 # so in this case, because we did not specify a specific
 # parameter, series within sites could differ by both
@@ -126,7 +134,7 @@ range(wdf_d$time)
 wdf_e <- aims_data("weather", api_key = NULL,
                    filters = list(parameter = "Chlorophyll",
                                   from_date = "2018-01-01",
-                                  thru_date = "2018-01-02"))$data
+                                  thru_date = "2018-01-02"))
[…]
