+++ man/capitalise.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/aimsdf-methods.R
+\name{capitalise}
+\alias{capitalise}
+\title{capitalise}
+\usage{
+capitalise(x)
+}
+\arguments{
+\item{x}{A character}
+}
+\description{
+Internal
+}
