+++ man/aimsdf-class.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/aimsdf-class.R
+\docType{class}
+\name{aimsdf-class}
+\alias{aimsdf-class}
+\alias{aimsdf}
+\title{Class \code{aimsdf} of data.frame downloaded by the \pkg{dataaimsr} package}
+\description{
+Datasets downloaded by the
+\code{\link[dataaimsr:dataaimsr-package]{dataaimsr}} package inherit
+the \code{aimsdf} class, which is data.frame with three attributes.
+}
+\details{
+See \code{methods(class = "aimsdf")} for an overview of available methods.
+}
+\seealso{
+\code{\link{aims_data}}
+}
