+++ man/aims_metadata.Rd
@@ -1,0 +1,27 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/aims_metadata.R
+\name{aims_metadata}
+\alias{aims_metadata}
+\title{Extracts metadata attribute from object of class aimsdf}
+\usage{
+aims_metadata(df_)
+}
+\arguments{
+\item{df_}{A data.frame of class \code{\link{aimsdf}} created by
+function \code{\link{aims_data}}}
+}
+\value{
+A \code{\link[base]{character}} vector.
+}
+\description{
+Extracts metadata attribute from object of class aimsdf
+}
+\details{
+This function retrieves the metadata attribute from an
+\code{\link{aimsdf}} object. If the input \code{\link{aimsdf}} object is
+a summary data.frame (see ?\code{\link{aims_data}}), then output will be
+an empty string.
+}
+\author{
+AIMS Datacentre \email{adc@aims.gov.au}
+}
