+++ man/make_pretty_data_label.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/aimsdf-methods.R
+\name{make_pretty_data_label}
+\alias{make_pretty_data_label}
+\title{make_pretty_data_label}
+\usage{
+make_pretty_data_label(x)
+}
+\arguments{
+\item{x}{A character}
+}
+\description{
+Internal
+}
