+++ man/make_pretty_colour.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/aimsdf-methods.R
+\name{make_pretty_colour}
+\alias{make_pretty_colour}
+\title{make_pretty_colour}
+\usage{
+make_pretty_colour(x, alpha_ = 0.55)
+}
+\arguments{
+\item{x}{A character}
+
+\item{alpha_}{A numeric}
+}
+\description{
+Internal
+}
