+++ man/is.aimsdf.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/aimsdf-class.R
+\name{is.aimsdf}
+\alias{is.aimsdf}
+\title{Checks if argument is a \code{aimsdf} object}
+\usage{
+is.aimsdf(x)
+}
+\arguments{
+\item{x}{An \R object}
+}
+\description{
+Checks if argument is a \code{aimsdf} object
+}

[…]
