### 2021-03-18 27aa4c7
moved API Key section inside the installation instruction
