### 2021-03-18 3f5c9b1
Added code of conduct and contributing files
