+++ R/aimsdf-methods.R
@@ -211,6 +211,7 @@ plot.aimsdf <- function(x, ..., ptype, pars) {
 #' Internal
 #'
 #' @param x A character
+#' @keywords internal
 make_pretty_data_label <- function(x) {
   ifelse(x == "weather", "Weather Station", "Temperature loggers")
 }
@@ -223,6 +224,7 @@ make_pretty_data_label <- function(x) {
 #' @param alpha_ A numeric
 #'
 #' @importFrom grDevices col2rgb rgb
+#' @keywords internal
 make_pretty_colour <- function(x, alpha_ = 0.55) {
   col <- col2rgb(x)
   rgb(col[1], col[2], col[3], alpha = alpha_ * 255, maxColorValue = 255)
@@ -233,6 +235,7 @@ make_pretty_colour <- function(x, alpha_ = 0.55) {
 #' Internal
 #'
 #' @param x A character
+#' @keywords internal
 capitalise <- function(x) {
   paste0(toupper(substr(x, 1, 1)), substr(x, 2, nchar(x)))
 }
@@ -243,6 +246,7 @@ capitalise <- function(x) {
 #'
 #' @param x An sfc_POINT
 #' @param ... Additional argument "pos" to internal function
+#' @keywords internal
 extract_map_coord <- function(x, ...) {
   sapply(x, function(z, pos)z[[pos]], ...)
 }
