+++ man/make_base_end_pt.Rd
@@ -22,3 +22,4 @@ via argument \code{summary} or \code{filters} in \code{\link{aims_data}}
 \author{
 AIMS Datacentre \email{adc@aims.gov.au}
 }
+\keyword{internal}
