+++ man/find_api_key.Rd
@@ -54,3 +54,4 @@ if it reaches the system's memory capacity.
 \author{
 AIMS Datacentre \email{adc@aims.gov.au}
 }
+\keyword{internal}
