+++ man/process_request.Rd
@@ -49,3 +49,4 @@ data request and processes result via
 \author{
 AIMS Datacentre \email{adc@aims.gov.au}
 }
+\keyword{internal}
