+++ man/is_aimsdf.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/aimsdf-class.R
-\name{is.aimsdf}
-\alias{is.aimsdf}
+\name{is_aimsdf}
+\alias{is_aimsdf}
 \title{Checks if argument is a \code{aimsdf} object}
 \usage{
-is.aimsdf(x)
+is_aimsdf(x)
 }
 \arguments{
 \item{x}{An \R object}
