### 2021-03-18 f2af2a3
added two examples for usage under README
