+++ README.Rmd
@@ -47,44 +47,14 @@ developed the [AIMS Data Platform API][1] which is a *REST API* providing
 JSON-formatted data to users. `dataaimsr` is an **R package** written to
 allow users to communicate with the AIMS Data Platform API using an API key
 and a few convenience functions to interrogate and understand the datasets
-that are available to download.
+that are available to download. In doing so, it allows the user to
+fully explore these datasets in R in whichever capacity they want (e.g.
+data visualisation, statistical analyses, etc). The package itself contains
+a `plot` method which allows the user to plot summaries of the different types
+of dataset made available by the API. Below we provide a brief context about
+the existing [Datasets][#datasets] that can be explored through `dataaimsr`.
 
 [1]: https://open-aims.github.io/data-platform/
-
-Currently, there are two AIMS long-term monitoring datasets available to be
-downloaded through `dataaimsr`:
-
-### Northern Australia Automated Marine Weather And Oceanographic Stations
-
-Automatic weather stations have been deployed by AIMS since 1980. Most of the
-stations are along the Great Barrier Reef (GBR) including the Torres Strait in
-North-Eastern Australia but there is also a station in Darwin and one at
-Ningaloo Reef in Western Australia. Many of the stations are located on the
-reef itself either on poles located in the reef lagoon or on tourist pontoons
-or other structures. A list of the weather stations which have been deployed
-by AIMS and the period of time for which data may be available can be
-found on the [metadata][2] webpage. **NB:** Records may not be continuous for
-the time spans given.
-
-[2]: https://apps.aims.gov.au/metadata/view/0887cb5b-b443-4e08-a169-038208109466
-
-### AIMS Sea Water Temperature Observing System (AIMS Temperature Logger Program)
-
-The data provided here are from a number of sea water temperature monitoring
-programs conducted in tropical and subtropical coral reefs environments around
-Australia. Data are available from approximately 80 GBR sites, 16 Coral Sea
-sites, 7 sites in North West Western Australia (WA), 8 Queensland regional
-ports, 13 sites in the Solitary Islands, 4 sites in Papua New Guinea and 10
-sites in the Cocos (Keeling) Islands. Data are obtained from in-situ data
-loggers deployed on the reef. Temperature instruments sample water
-temperatures every 5-10 minutes (typically) and are exchanged and downloaded
-approximately every 12 months. Temperature loggers on the reef-flat are
-generally placed just below Lowest Astronomical Tide level. Reef-slope (or
-where specified as Upper reef-slope) generally refers to depths 5--9 m while
-Deep reef-slope refers to depths of ~20 m. For more information on the dataset
-and its usage, please visit the [metadata][3] webpage.
-
-[3]: https://apps.aims.gov.au/metadata/view/4a12a8c0-c573-11dc-b99b-00008a07204e 
 
 ## Installation
 
@@ -161,6 +131,43 @@ More comprehensive examples about how to navigate `dataaimsr` and interrogate
 the datasets can be found on our [online vignettes][4].
 
 [4]: https://open-AIMS.github.io/dataaimsr/articles
+
+## Datasets
+
+Currently, there are two AIMS long-term monitoring datasets available to be
+downloaded through `dataaimsr`:
+
+### Northern Australia Automated Marine Weather And Oceanographic Stations
+
+Automatic weather stations have been deployed by AIMS since 1980. Most of the
+stations are along the Great Barrier Reef (GBR) including the Torres Strait in
+North-Eastern Australia but there is also a station in Darwin and one at
+Ningaloo Reef in Western Australia. Many of the stations are located on the
+reef itself either on poles located in the reef lagoon or on tourist pontoons
[…]
