### 2021-03-18 aebcb13
created mock env variable to fake no internet connection in tests
