### 2021-03-18 81207a7
make convenience functions fail if input is not aimsdf; added accompanying tests
