### 2021-03-18 84104c7
sent helpers from methods to helpers files, updated test, added tests for methods
