+++ man/make_pretty_data_label.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/aimsdf-methods.R
+% Please edit documentation in R/helpers.R
 \name{make_pretty_data_label}
 \alias{make_pretty_data_label}
 \title{make_pretty_data_label}
