+++ man/extract_map_coord.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/aimsdf-methods.R
+% Please edit documentation in R/helpers.R
 \name{extract_map_coord}
 \alias{extract_map_coord}
 \title{extract_map_coord}
