+++ man/capitalise.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/aimsdf-methods.R
+% Please edit documentation in R/helpers.R
 \name{capitalise}
 \alias{capitalise}
 \title{capitalise}
