### 2021-03-18 2b5cd66
updated paper text and map
