+++ paper/paper.bib
@@ -1,3 +1,26 @@
+@article{morelo2014methods,
+  author = {Morello, E. B. and Galibert, G. and Smith, D. and Ridgway, K. R. and Howell, B. and Slawinski, D. and Timms, G. P. and Evans, K. and Lynch, T. P.},
+  title = {{Quality Control (QC)} procedures for {Australia's National Reference Station's} sensor data—{Comparing} semi-autonomous systems to an expert oceanographer},
+  year = {2014},
+  journal = {Methods in Oceanography},
+  volume = {9},
+  pages = {17-33},
+  url = {https://doi.org/10.1016/j.mio.2014.09.001},
+  doi = {10.1016/j.mio.2014.09.001}
+}
+
+@article{toma2016acta,
+  author = {Toma, D. M. and Benadí, A. G. and {Manuel-Gonzalez}, B. J. and {del-Río-Fernandez}, J.},
+  title = {Systematic quality control for long term ocean observations and applications},
+  year = {2016},
+  journal = {Acta Imeko},
+  volume = {5},
+  issue = {1},
+  pages = {64-68},
+  url = {https://doi.org/10.21014/acta_imeko.v5i1.213},
+  doi = {10.21014/acta_imeko.v5i1.213}
+}
+
 @manual{rcitation,
   title = {R: A Language and Environment for Statistical Computing},
   author = {{R Core Team}},
@@ -7,7 +30,7 @@
   url = {https://www.R-project.org/},
 }
 
-@manual{httr,
+@manual{httrcit,
   title = {httr: Tools for Working with URLs and HTTP},
   author = {Hadley Wickham},
   year = {2020},
@@ -15,7 +38,7 @@
   url = {https://CRAN.R-project.org/package=httr},
 }
 
-@article{jsonlite,
+@article{jsonlitecit,
   title = {The jsonlite Package: A Practical and Consistent Mapping Between JSON Data and R Objects},
   author = {Jeroen Ooms},
   journal = {arXiv:1403.2805 [stat.CO]},
@@ -23,10 +46,72 @@
   url = {https://arxiv.org/abs/1403.2805},
 }
 
-@manual{parsedate,
+@manual{parsedatecit,
   title = {parsedate: Recognize and Parse Dates in Various Formats, Including All ISO 8601 Formats},
   author = {Gábor Csárdi and Linus Torvalds},
   year = {2019},
   note = {R package version 1.2.0},
   url = {https://CRAN.R-project.org/package=parsedate},
+}
+
+@Manual{dplyrcit,
+  title = {dplyr: A Grammar of Data Manipulation},
+  author = {Hadley Wickham and Romain François and Lionel Henry and Kirill Müller},
+  year = {2021},
+  note = {R package version 1.0.5},
+  url = {https://CRAN.R-project.org/package=dplyr},
+}
+
+@Manual{tidyrcit,
+  title = {tidyr: Tidy Messy Data},
+  author = {Hadley Wickham},
+  year = {2021},
+  note = {R package version 1.1.3},
+  url = {https://CRAN.R-project.org/package=tidyr},
+}
+
+@Manual{rnaturalearthcit,
+  title = {rnaturalearth: World Map Data from Natural Earth},
+  author = {Andy South},
+  year = {2017},
+  note = {R package version 0.1.0},
+  url = {https://CRAN.R-project.org/package=rnaturalearth},
+}
+
+@Article{sfcit,
+  author = {Edzer Pebesma},
+  title = {{Simple Features for R: Standardized Support for Spatial Vector Data}},
+  year = {2018},
+  journal = {{The R Journal}},
+  doi = {10.32614/RJ-2018-009},
[…]
