+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'dataaimsr: A Client for the Australian Institute of Marine Science Data Platform API which provides easy access to AIMS Data Platform'
+title: 'dataaimsr: An R Client for the Australian Institute of Marine Science Data Platform API which provides easy access to AIMS Data Platform'
 tags:
   - R
   - sea surface temperature
@@ -13,9 +13,23 @@ authors:
     affiliation: "1, 2"
   - name: Greg Coleman
     affiliation: "3"
+  - name: Duncan Fermor
+    affiliation: "3"
+  - name: Eduardo Klein
+    affiliation: "3"
+  - name: Tobias Robinson
+    affiliation: "3"
+  - name: Jason Smith
+    affiliation: "3"
   - name: Jeffrey L. Sheehan
     affiliation: "3"
-  - name: Duncan Fermor
+  - name: Shannon Dowley
+    affiliation: "3"
+  - name: Dean Ditton
+    affiliation: "3"
+  - name: Kevin Gunn
+    affiliation: "3"
+  - name: Gavin Ericson
     affiliation: "3"
   - name: Murray Logan
     affiliation: "3"
@@ -30,7 +44,7 @@ affiliations:
    index: 3
 
 citation_author: Barneche et al.
-date: "2021-02-11"
+date: "2021-03-18"
 bibliography: paper.bib
 output:
   my_modified_joss:
@@ -61,7 +75,11 @@ developed the [AIMS Data Platform API][1] which is a *REST API* providing
 JSON-formatted data to users. `dataaimsr` is an **R package** written to
 allow users to communicate with the AIMS Data Platform API using an API key
 and a few convenience functions to interrogate and understand the datasets
-that are available to download.
+that are available to download. In doing so, it allows the user to
+fully explore these datasets in R in whichever capacity they want (e.g.
+data visualisation, statistical analyses, etc). The package itself contains
+a `plot` method which allows the user to plot summaries of the different types
+of dataset made available by the API.
 
 [1]: https://open-aims.github.io/data-platform/
 
@@ -103,9 +121,17 @@ and its usage, please visit the [AIMS metadata][3] webpage.
 # Technical details and Usage
 
 Before loading the package, a user needs to download and store their personal
-[AIMS Data Platform API Key][4]---we recommend it storing it in `.Renviron`.
+[AIMS Data Platform API Key][4]---we strongly encourage users to
+maintain their API key as a private, locally hidden environment variable
+(`AIMS_DATAPLATFORM_API_KEY`) in the `.Renviron` file for
+automatic loading at the start of an R session.
 
 [4]: https://open-aims.github.io/data-platform/key-request
+
+`dataaimsr` imports the packages *httr* [@httrcit], *jsonlite* [@jsonlitecit],
+*parsedate* [@parsedatecit], *dplyr* [@dplyrcit], *tidyr* [@tidyrcit],
+*rnaturalearth* [@rnaturalearthcit], *sf* [@sfcit], *ggplot2* [@ggplot2cit],
+*ggrepel* [@ggrepelcit] and *curl* [@curlcit].
 
 The [Weather Station][2] and [Sea Water Temperature Loggers][3] datasets are
 very large (terabytes in size), and as such they are not locally stored.
@@ -136,29 +162,26 @@ download data.
 ### Data summaries
 
 The first step would be to visualise the dataset. We do this by
-mapping all available sites. First we need to store the DOI for the
-target dataset---this is done via the function `aims_data_doi`. We
-can then download the summary information for the Sea Water Temperature 
-Loggers dataset using the main function called `aims_data`:
+mapping all available sites. For example, we download the summary information
+for the Sea Water Temperature Loggers dataset using the main function called
+`aims_data`:
 
 
 ```r
[…]
