### 2021-03-18 15974b4
added vignette precompilation trick suggested by @boshek
