### 2021-03-18 f252bdd
implemented tests with httptest as suggested by @techisdead
