+++ tests/testthat/api.aims.gov.au/data-v2.0/10.25845/5c09bf93f315d/data/summary-by-series.R
@@ -1,0 +1,21 @@
+structure(list(url = "https://api.aims.gov.au/data-v2.0/10.25845/5c09bf93f315d/data/summary-by-series", 
+    status_code = 403L, headers = structure(list(date = "Thu, 18 Mar 2021 10:16:10 GMT", 
+        `content-type` = "application/json", `content-length` = "42", 
+        `x-amzn-requestid` = "4d972c44-c5bf-4d90-9d6f-975ef875e806", 
+        `x-amzn-errortype` = "MissingAuthenticationTokenException", 
+        `x-amz-apigw-id` = "cYNAlEuISwMFrXQ="), class = c("insensitive", 
+    "list")), all_headers = list(list(status = 403L, version = "HTTP/2", 
+        headers = structure(list(date = "Thu, 18 Mar 2021 10:16:10 GMT", 
+            `content-type` = "application/json", `content-length` = "42", 
+            `x-amzn-requestid` = "4d972c44-c5bf-4d90-9d6f-975ef875e806", 
+            `x-amzn-errortype` = "MissingAuthenticationTokenException", 
+            `x-amz-apigw-id` = "cYNAlEuISwMFrXQ="), class = c("insensitive", 
+        "list")))), cookies = structure(list(domain = logical(0), 
+        flag = logical(0), path = logical(0), secure = logical(0), 
+        expiration = structure(numeric(0), class = c("POSIXct", 
+        "POSIXt")), name = logical(0), value = logical(0)), row.names = integer(0), class = "data.frame"), 
+    content = charToRaw("{\"message\":\"Missing Authentication Token\"}"), 
+    date = structure(1616062570, class = c("POSIXct", "POSIXt"
+    ), tzone = "GMT"), times = c(redirect = 0, namelookup = 4.8e-05, 
+    connect = 4.9e-05, pretransfer = 9.6e-05, starttransfer = 0.063845, 
+    total = 0.06386)), class = "response")

[…]
