### 2021-03-18 d04dde2
bumped version and updated NEWS
