+++ .github/workflows/test-coverage.yaml
@@ -39,7 +39,7 @@ jobs:
       - name: Install dependencies
         run: |
           install.packages(c("remotes"))
-          remotes::install_deps(dependencies = TRUE)
+          remotes::install_deps(dependencies = TRUE, type = "binary")
           remotes::install_cran("covr")
         shell: Rscript {0}
 
