+++ .github/workflows/pkgdown.yaml
@@ -27,8 +27,9 @@ jobs:
 
       - name: Install dependencies
         run: |
-          install.packages(c("ps", "remotes", "tidyverse", "anytime"))
-          remotes::install_deps(dependencies = TRUE)
+          install.packages(c("ps", "remotes", "tidyverse", "anytime", "sf",
+                             "units"), type = "binary")
+          remotes::install_deps(dependencies = TRUE, type = "binary")
           remotes::install_dev("pkgdown")
         shell: Rscript {0}
 
