+++ pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 2.7.3
 pkgdown: 1.6.1.9000
-pkgdown_sha: 1c28ef9d9476e35974e1933de76a9aaa24d4cd27
+pkgdown_sha: ae4332d48aedc46be2b532b26661a9bc6b7244fc
 articles:
   navigating: navigating.html
   tloggers: tloggers.html
   weather: weather.html
-last_built: 2021-02-16T00:52Z
+last_built: 2021-03-18T10:51Z
 
