+++ .github/workflows/R-CMD-check.yaml
@@ -53,18 +53,30 @@ jobs:
           Rscript -e "remotes::install_github('r-hub/sysreqs')"
           sysreqs=$(Rscript -e "cat(sysreqs::sysreq_commands('DESCRIPTION'))")
           sudo -s eval "$sysreqs"
+          sudo apt-get install gdal-bin proj-bin libgdal-dev libproj-dev
 
-      - name: Install dependencies
+      - name: Install dependencies OS_Win
+        if: runner.os != 'Linux'
         run: |
           install.packages(c("ps", "remotes", "tidyverse", "anytime", "sf",
                              "units"), type = "binary")
           remotes::install_deps(dependencies = TRUE, type = "binary")
         shell: Rscript {0}
 
+      - name: Install dependencies Linux
+        if: runner.os == 'Linux'
+        run: |
+          install.packages(c("ps", "remotes", "tidyverse", "anytime", "sf",
+                             "units"))
+          remotes::install_deps(dependencies = TRUE, type = "binary")
+        shell: Rscript {0}
+
       - name: Check
         env:
           _R_CHECK_CRAN_INCOMING_REMOTE_: false
-        run: rcmdcheck::rcmdcheck(args = c("--no-manual", "--as-cran"), error_on = "warning", check_dir = "check")
+        run: |
+          install.packages("rcmdcheck")
+          rcmdcheck::rcmdcheck(args = c("--no-manual", "--as-cran"), error_on = "warning", check_dir = "check")
         shell: Rscript {0}
 
       - name: Upload check results
