+++ pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 2.7.3
 pkgdown: 1.6.1.9000
-pkgdown_sha: ae4332d48aedc46be2b532b26661a9bc6b7244fc
+pkgdown_sha: 1ec304b0002d0483f386e5a9b66cc0cde24b335f
 articles:
   navigating: navigating.html
   tloggers: tloggers.html
   weather: weather.html
-last_built: 2021-03-18T10:51Z
+last_built: 2021-03-19T00:59Z
 
