### 2021-03-19 #428 @dbarneche
I would like to thank the editor @ldecicco-USGS and both referees @boshek and @techisdead for dedicating their time to reviewing our package. I have tried my best to incorporate all suggestions raised. Changes have been summarised in [NEWS.md](https://github.com/open-AIMS/dataaimsr/blob/master/NEWS.md); below I provide responses with the corresponding commit links. I look forward to your response.
Best wishes,
Diego

## In response to suggestions made by @boshek:

> @ldecicco-USGS here is my review of the `dataaimsr`. Overall a really great package that accesses a valuable data source. I have tried to summarize my key topics below. I think the biggest two point for me are:

> 1. Spend some additional time to make sure the API keys and the presence of an internet connections are checked for both users and for CI/CRAN services. This will help you debug future issues and provides some nice guardrails for users.
> 2. I'd advocate strongly that `aims_data` reliably returns a data.frame and exposes those other elements (citation etc) as attributes.
>
> Great job with this package!

**Response:** Thank you very much @boshek. I have tried my best to incorporate almost all of your excellent suggestions. There were only two instances where I diverged in opinion, below I provide a point-by-point response to each comment.

> ## Package Review
> * [x]  As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).
>
> #### Documentation
> The package includes all the following forms of documentation:
>
> * [x]  **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
> * [ ]  **Installation instructions:** for the development version of package and any non-standard dependencies in README
> * [x]  **Vignette(s)** demonstrating major functionality that runs successfully locally
> * [ ]  **Function Documentation:** for all exported functions
> * [ ]  **Examples** (that run successfully locally) for all exported functions
> * [ ]  **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).
>
> > ##### For packages co-submitting to JOSS
> >
> > * [x]  The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
> >
> > The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
> >
> > * [x]  **A short summary** describing the high-level functionality of the software
> > * [x]  **Authors:**  A list of authors with their affiliations
> > * [x]  **A statement of need** clearly stating problems the software is designed to solve and its target audience.
> > * [ ]  **References:** with DOIs for all those that have one (e.g. papers, datasets, software).
>
> #### Functionality
> * [x]  **Installation:** Installation succeeds as documented.
> * [x]  **Functionality:** Any functional claims of the software been confirmed.
> * [x]  **Performance:** Any performance claims of the software been confirmed.
> * [x]  **Automated tests:** Unit tests cover essential functions of the package
>   and a reasonable range of inputs and conditions. All tests pass on the local machine.
> * [x]  **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines
>
> Estimated hours spent reviewing: 8.
>
> * [x]  Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

**Response:** You are now officially listed as a package reviewer together with @ldecicco-USGS and @techisdead. Change implemented in [open-aims/dataaimsr@`c4a26fb`](https://github.com/open-AIMS/dataaimsr/commit/c4a26fb50312cd8a74be205bf3a4689a0dd7ba04).

> # Review Comments
> ## Documentation
> * **Statement of Need**: I think that the README could actually be made _shorter_. I think it would be helpful to outline why a user would want to use R to interact with this data. To me the specific value add here is that you are connecting the AIMS api with all the amazing capabilities of R. I think the text in the _Overview_ section is valuable. I just wonder if it could be condensed a little to get users to specifically to the installation instructions as really what they are after here is the software itself.

**Response:** I appreciate this suggestion. I have expanded the overview section by a few sentences to better highlight your comment. However, in case it's still OK with you, I'd prefer to maintain the blurbs about the datasets on the README. That is because I suspect that many users do not even know about the existence of these datasets, and so having a summary about them should help break that barrier. That might have been why @techisdead appreciated the longer README? I do take your point though that users may be looking for installation instructions above anything else, so I have moved those two paragraphs to the end of README, after the installation instructions and usage examples. I hope this serves as a good compromise, but please do let me know in case you feel strongly about this. Changes implemented in [open-aims/dataaimsr@`11dbdf2`](https://github.com/open-AIMS/dataaimsr/commit/11dbdf257a4b8545925ebae187d99e41d06c39f8).

> * **Installation instructions**: I would revamp the installation instructions so that getting the API key is explictly a step in the installation instructions. I would also include information about how the API key should be kept secret. You might consider linking to [this](https://cran.r-project.org/web/packages/httr/vignettes/secrets.html) article.

**Response:** Excellent suggestion. Incorporated in [open-aims/dataaimsr@`27aa4c7`](https://github.com/open-AIMS/dataaimsr/commit/27aa4c7b1cac2c7a754a3ef35d88b15dfff88204).

> * **Vignette(s)**: See comment in API keys section. I don't think ggmap is a particularly useful here. Given ggmaps' own API requirements I would strongly advise against including it here as a dependency and including it in the vignettes. If you really want a map, I would suggest try an sf approach. But I think there are plenty of tutorials around to tackle mapping in R and it detracts from the core focus of dataaimsr.

**Response:** Thanks for this suggestion. I built on the example code you kindly provided below for the maps to create a plot method (please see more about this below). The consequence of that change though is that `ggmap` is no longer a requirement for the vignettes nor a dependency of the package. Change implemented in [open-aims/dataaimsr@`557f41a`](https://github.com/open-AIMS/dataaimsr/commit/557f41a10f42af22217d9b490bca5d97840ab532).

> * **Examples**: I think the example in `aims_data` could be significantly shortened. The examples section is meant to be illustrative rather than comprehensive. But I wouldn't waste all that great code. This is an opinion so not a requirement for acceptance.  I would simply move it to a vignette. I am a bit uncertain how you are going to manage with CRAN, the requirement for the API keys and the ability to run examples. Here I think you might need to build in some ability for the function to 'fail gracefully' when the API key isn't available. That is, don't error but instead return an informative message. That same applies to checking if someone has a) internet and b) if the API is down for whatever reason. Again this is particularly relevant if CRAN submission is desired. I think you can wrap them in a `try` function. That might work.

**Response:** Thanks for this suggestion. I hope this is OK, but I would prefer to keep the extended list of examples. I think many people might be a bit lost when first using the package---for the same reasons raised in the comment RE: REAMDE above---so those examples should help them better understand the capabilities of the package. As for the failures, thank you! I have implemented two steps to fail the function gracefully. One is related to not having internet connection (as per your suggestion below using `curl::has_internet()`), and the other happens in case the input filter values are incorrect. Both implemented in [open-aims/dataaimsr@`3ead676`](https://github.com/open-AIMS/dataaimsr/commit/3ead67654944ffcdad967561dffde197ebc6d898).

I'm unsure though if I should also implement something along these lines for
the internal function `handle_error`?

>   **Community guidelines**: Missing a CONTRIBUTING guide _and_ (though not required) a CODE_OF_CONDUCT file.

**Response:** Done in [open-aims/dataaimsr@`3f5c9b1`](https://github.com/open-AIMS/dataaimsr/commit/3f5c9b16e75c6813ea1c7ce0a4001de2e4736e8b).

> ## Paper
> * I think it is worth adopting stronger language about the APi keys in the paper to ensure that people don't expose those publicly.

**Response:** Agreed 100%. Changes implemented in [open-aims/dataaimsr@`27aa4c7`](https://github.com/open-AIMS/dataaimsr/commit/27aa4c7b1cac2c7a754a3ef35d88b15dfff88204) for the `README.md` section, and in [open-aims/dataaimsr@`557f41a`](https://github.com/open-AIMS/dataaimsr/commit/557f41a10f42af22217d9b490bca5d97840ab532) for the vignettes.

> * I would modify the map presented in the paper. This isn't a scientific map. Either I would remove it or I would provide something that use open data. Something like this roughed up example would work:
>
> ```r
> library(dataaimsr)
> library(sf)
> library(rnaturalearth)
> library(tmap)
>
> # see ?aims_data_doi for argument names
> ssts_doi <- aims_data_doi("temp_loggers")
> sdata <- aims_data(ssts_doi, api_key = NULL,
>                    summary = "summary-by-series")
>
> oceania_boundary <- ne_countries(continent = "oceania", returnclass = "sf") %>%
>   st_transform(3112)
>
> sdata_sf <- sdata %>%
>   dplyr::mutate(cols = cal_obs * 1e-3) %>%
>   tidyr::drop_na(lon, lat) %>%
>   st_as_sf(coords = c("lon", "lat"), crs = 4326) %>%
>   st_transform(3112)
>
> tm_shape(oceania_boundary) +
>   tm_polygons() +
>   tm_shape(sdata_sf) +
>   tm_dots(col = 'cols', size = 0.1)
> ```

**Response:** Thank you very much for your suggestions. I hope you don't mind, but I took the liberty to build on the above code not only to change the paper but also to create a plot method for the new class (more on this below). We have also 1) updated the list of authors, 2) added a blurb about the sister web visualisation tool, and 3) added references and text for the quality control calibration routine. Changes implemented in [open-aims/dataaimsr@`2b5cd66`](https://github.com/open-AIMS/dataaimsr/commit/2b5cd66d2c0954b10d8e39b9f0fd43cb825ac7a5).

> ## General
> * I think that your R dependency could be 3.3.0 rather than 3.6.0.

**Response:** This dependency has been updated in [open-aims/dataaimsr@`c4a26fb`](https://github.com/open-AIMS/dataaimsr/commit/c4a26fb50312cd8a74be205bf3a4689a0dd7ba04).

> * I am wondering about the usefulness of the `aims_data_doi` function. Would it not be more straight forward for a user to simply to expose the `target` argument directly to `aims_data` and other functions like `expose_attributes`?

**Response:** I agree that `aims_data_doi` was not that useful. Your suggested changes have been fully incorporated in [open-aims/dataaimsr@`a837afd`](https://github.com/open-AIMS/dataaimsr/commit/a837afdc49626d63305eba43924b8d5a950c00f8).

> * I would suggest that adding `aims_*` prefixes to all exposed functions would be a good idea.

**Response:** Done in [open-aims/dataaimsr@`ec8f85c`](https://github.com/open-AIMS/dataaimsr/commit/ec8f85cc1b52556e233067e1bb5951b6cf4af05f).

> * I think you reconsider exposing most of the functions that handle http requests. That would things like `handle_error` and `page_data`. Those all can be internal functions and don't need to be user facing.

**Response:** Perfect. Now the only exported functions are:

```
> library(dataaimsr)
> dataaimsr::
dataaimsr::aims_citation           dataaimsr::aims_expose_attributes  dataaimsr::aims_metadata           dataaimsr::is_aimsdf
dataaimsr::aims_data               dataaimsr::aims_filter_values      dataaimsr::aims_parameters
```

Change implemented in [open-aims/dataaimsr@`d32bc86`](https://github.com/open-AIMS/dataaimsr/commit/d32bc86b69337f5614fcd6c648ddc5c5f327d323). I still don't know how to prevent the non-exported functions from appearing in the [pkgdown reference page](https://open-aims.github.io/dataaimsr/reference/index.html)---any ideas?

> * If the object returned by `aims_data` was always just a data.frame I think that would be beneficial. Instead of returning a list I think you could stored those other elements as attributes attached to the object and then construct extract functions. So something like `aims_citation` which takes the object returned by `aims_data` and return the citation. Typically returning objects of different classes depending on the value of an argument is considered not best practice.
> * I think this approach could be extended to parameters
> * I don't consider this required for acceptance but I would highly recommend considering adding classes to the `aims_data` object and implementing `print` and `plot` methods.

**Response:** I hope you don't mind me amalgamating all these excellent suggestions for a single answer, but basically I have implemented all of them. `aims_data` now returns a `data.frame` of class `aimsdf` with 5 attributes, 3 of which can be extracted using the convenience functions `aims_citation`, `aims_metadata` and `aims_parameters`. The other two attributes, `type` and `target` were added to facilitate plotting methods. `plot`, `summary`, `print` have been extended to work with class `aimsdf`. The `plot` can be either a time series or a map. These changes were implemented in [open-aims/dataaimsr@`3ead676`](https://github.com/open-AIMS/dataaimsr/commit/3ead67654944ffcdad967561dffde197ebc6d898), [open-aims/dataaimsr@`799e4c2`](https://github.com/open-AIMS/dataaimsr/commit/799e4c2d71145082f998d78d6677bad943422e54) and [open-aims/dataaimsr@`1330e0a`](https://github.com/open-AIMS/dataaimsr/commit/1330e0a378de430db545f9e17595d2b8051c9408).

> * Add something like `curl::has_internet()` to fail "gracefully" when lacking an internet connection.

**Response:** As mentioned above, done in [open-aims/dataaimsr@`3ead676`](https://github.com/open-AIMS/dataaimsr/commit/3ead67654944ffcdad967561dffde197ebc6d898).

> * I think the README could benefit from a short usage example.

**Response:** Done in [open-aims/dataaimsr@`f2af2a3`](https://github.com/open-AIMS/dataaimsr/commit/f2af2a354a399c42734657083078b97dca7f9d99).

> * In the DESCRIPTION I think the url could also include a link to the API. They just need to be separated by a comma.

**Response:** Done in [open-aims/dataaimsr@`c4a26fb`](https://github.com/open-AIMS/dataaimsr/commit/c4a26fb50312cd8a74be205bf3a4689a0dd7ba04).

> ## API keys
> * I think that it would be worthwhile in include a check for the env variable when the package is loaded. Using `find_api_key` would probably work fine.
> * Exclude the actual API from the RMD's. There are few ways around actually needing to include it. You could precompile your vignettes like this:
>
>   * https://github.com/bcgov/bcdata/blob/master/vignettes/precompile.R
>   * Or you can set an phantom environmental variable which results in nothing actually being run on CRAN: https://github.com/ropensci/tidyhydat/blob/2c177aec31d5e8fe296b5336b786b01bb83841c9/vignettes/tidyhydat_an_introduction.Rmd#L18

**Response:** Thanks for these examples. I decided to implement the pre-compilation trick suggested. I would appreciate it if you could confirm whether the way I adapted the original `precompile.R` file is properly carrying over the original license. Changes in [open-aims/dataaimsr@`15974b4`](https://github.com/open-AIMS/dataaimsr/commit/15974b4a328f29ec58df2fb4993920290704b07e).

> * For testing, I think the best approach if you want to test directly on the API is to add a token as GitHub secret. However ropensci does have a suite mocking tools that facilitate exactly this type of thing. For example: https://github.com/ropensci/webmockr

**Response:** Thanks for raising this issue. I have also removed the key from the test suite. I decided to adopt the `httptest` mock approach as suggested below by @techisdead. Changes in [open-aims/dataaimsr@`f252bdd`](https://github.com/open-AIMS/dataaimsr/commit/f252bdd500cf19714e3094882ec14ea076667b60).

## In response to suggestions made by @techisdead:

> ## Package Review
> _Please check off boxes as applicable, and elaborate in comments below. Your review is not limited to these topics, as described in the reviewer guide_
>
> * **Briefly describe any working relationship you have (had) with the package authors.**
> * [x]  As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).
>
> #### Documentation
> The package includes all the following forms of documentation:
>
> * [x]  **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
> * [x]  **Installation instructions:** for the development version of package and any non-standard dependencies in README
> * [x]  **Vignette(s)** demonstrating major functionality that runs successfully locally
> * [x]  **Function Documentation:** for all exported functions
> * [x]  **Examples** (that run successfully locally) for all exported functions
> * [x]  **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).
>
> > ##### For packages co-submitting to JOSS
> >
> > * [x]  The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
> >
> > The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
> >
> > * [x]  **A short summary** describing the high-level functionality of the software
> > * [x]  **Authors:**  A list of authors with their affiliations
> > * [x]  **A statement of need** clearly stating problems the software is designed to solve and its target audience.
> > * [ ]  **References:** with DOIs for all those that have one (e.g. papers, datasets, software).
>
> #### Functionality
> * [x]  **Installation:** Installation succeeds as documented.
> * [x]  **Functionality:** Any functional claims of the software been confirmed.
> * [x]  **Performance:** Any performance claims of the software been confirmed.
> * [ ]  **Automated tests:** Unit tests cover essential functions of the package
>   and a reasonable range of inputs and conditions. All tests pass on the local machine.
> * [x]  **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines
>
> Estimated hours spent reviewing: 4
>
> * [x]  Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

**Response:** Thank you very much your review @techisdead. You are now officially listed as a package reviewer together with @ldecicco-USGS and @boshek. Change implemented in [open-aims/dataaimsr@`c4a26fb`](https://github.com/open-AIMS/dataaimsr/commit/c4a26fb50312cd8a74be205bf3a4689a0dd7ba04).

> ### Review Comments
> Great package and data sets. I'm excited to see how the API evolves and think it's great this package is being developed alongside.
>
[…]
