### 2021-03-19 371f499
removed binary requirement for installation across OS, adding external libraries for Mac and Winodws, as well as rgeos and rgdal as package dependencies for CI to work
