+++ pkgdown.yml
@@ -5,5 +5,5 @@ articles:
   navigating: navigating.html
   tloggers: tloggers.html
   weather: weather.html
-last_built: 2021-03-19T00:59Z
+last_built: 2021-03-19T01:36Z
 
