### 2021-03-19 93c3cf5
removed dplyr pipe from paper Rmd
