### 2021-03-19 3b1e4f8
fixing namespace issues + attempt to make cmd check work again
