+++ man/head.aimsdf.Rd
@@ -1,17 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/aimsdf-methods.R
-\name{head.aimsdf}
-\alias{head.aimsdf}
-\title{head.aimsdf}
-\usage{
-\method{head}{aimsdf}(x, ...)
-}
-\arguments{
-\item{x}{An object of class \code{\link{aimsdf}} as
-returned by \code{\link{aims_data}}.}
-
-\item{...}{Not used.}
-}
-\description{
-head.aimsdf
-}
