+++ NAMESPACE
@@ -1,10 +1,8 @@
 # Generated by roxygen2: do not edit by hand
 
-S3method(head,aimsdf)
 S3method(plot,aimsdf)
 S3method(print,aimsdf)
 S3method(summary,aimsdf)
-S3method(tail,aimsdf)
 export(aims_citation)
 export(aims_data)
 export(aims_expose_attributes)
@@ -47,3 +45,4 @@ importFrom(sf,st_as_sf)
 importFrom(sf,st_transform)
 importFrom(tidyr,drop_na)
 importFrom(utils,URLencode)
+importFrom(utils,head)
