### 2021-03-19 7fc3518
borrowed part of r cmd check actions from sf github page
