### 2021-03-19 1e38d79
more attempts to fix sf dependencies
