### 2021-03-20 #428 @techisdead
I've had another look at the updates - Thanks for the organised response and all the work to address our comments.  I'm happy that it's all been addressed.  The vignettes all work fine, updating the tests avoids exposing and API key and all the functions work as per the vignettes and manuals (unless I've missed something).

I only found one remaining niggle - Tests failed (when invoked from `devtools::check()` but not `devtools:test()` which usually points to a namespace issue).   This was the error

```
> test_check("dataaimsr")
── Error (test-aimsdf-methods.R:38:5): Correct structure ───────────────────────
Error: package rgeos required for finding out which hole belongs to which exterior ring
Backtrace:
     █
  1. ├─base::plot(wa, ptype = "map") test-aimsdf-methods.R:38:4
  2. ├─dataaimsr:::plot.aimsdf(wa, ptype = "map")
  3. │ ├─ne_countries(continent = "oceania", returnclass = "sf") %>% st_transform(crs = 3112)
  4. │ └─rnaturalearth::ne_countries(continent = "oceania", returnclass = "sf")
  5. │   └─rnaturalearth:::ne_as_sf(spdf[filter, ], returnclass)
  6. │     ├─sf::st_as_sf(x)
  7. │     └─sf:::st_as_sf.Spatial(x)
  8. │       ├─sf::st_as_sfc(sp::geometry(x), ...)
  9. │       └─sf:::st_as_sfc.SpatialPolygons(sp::geometry(x), ...)
 10. └─sf::st_transform(., crs = 3112)

```

The fix for me was adding `rgeos` to 'Suggests' in the DESCRIPTION file.

Thanks again for this package!
