### 2021-03-22 #428 @boshek
Great responses! This looks great to me. With respect to the README that is certainly fine.

Re pkgdown issue:

> I still don't know how to prevent the non-exported functions from appearing in the pkgdown reference page---any ideas?

This is because they are still documented. You can either remove the documentation or use `@noRd`. But then you'll also need to remove references to that function.

This package looks to be in great shape now. Well done @dbarneche!
