+++ DESCRIPTION
@@ -28,7 +28,7 @@ License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
 VignetteBuilder: knitr
-URL: https://open-aims.github.io/dataaimsr, https://open-aims.github.io/data-platform/key-request, https://open-aims.github.io/data-platform/swagger/
-BugReports: https://github.com/open-aims/dataaimsr/issues
+URL: https://ropensci.github.io/dataaimsr, https://open-aims.github.io/data-platform/key-request, https://open-aims.github.io/data-platform/swagger/
+BugReports: https://github.com/ropensci/dataaimsr/issues
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.1.1
