### 2021-03-29 #428 @dbarneche
Hi @ldecicco-USGS , I'm almost done with the transfer. I'm just confused about the re-directing part to create the ropensci docs. Previously the package was hosted under our GitHub organisation account `open-AIMS`. I then had a look here:
https://devguide.ropensci.org/redirect.html
But wasn't really clear on how setting up and `open-AIMS.github.io` repo with a `dataaimsr` folder would help redirect things given that everything now has been transferred to `ropensci`? Would it be possible to maintain the current pkgdown GitHub Action *and* have ropensci docs?
Sorry about the confusion, and thanks for your guidance!
