+++ pkgdown.yml
@@ -5,5 +5,5 @@ articles:
   navigating: navigating.html
   tloggers: tloggers.html
   weather: weather.html
-last_built: 2021-03-29T01:11Z
+last_built: 2021-03-29T01:33Z
 
