### 2021-03-30 #428 @dbarneche
Oh, I see. Thanks for clarifying!
Currently pkgdown website is hosted under
https://ropensci.github.io/dataaimsr/

I'm assuming that's OK?

Also, how should I proceed with submission to JOSS? Is there a path I should take from rOpenSci? Or do I proceed normally as a regular submission through their system?
