### 2021-03-31 #428 @boshek
One thing @dbarneche to mention. The rOpenSci development server generates nightly builds of your package. So this should work now too:

```
install.packages("dataaimsr ", repos = "https://dev.ropensci.org")
```
So you can include that in your README. Here is a good example: https://github.com/ropensci/targets#installation
