### 2021-03-31 #428 @ldecicco-USGS
You should proceed with the JOSS submission in the regular way. Somewhere in the submission add a link to this thread though, that will be helpful.
